feat: add a script that will update markdown instead
