--- folders.json
@@ -1,7 +1,0 @@
-{
-  "folders": [
-    "1-template",
-    "2-shapes",
-    "3-materials"
-  ]
-}
