--- index.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="style.css">
-  <script src="/main.js" defer></script>
-  <title>Three.js Demos</title>
-</head>
-
-<body>
-  <h1>
-    Three.js Demos
-  </h1>
-  <main>
-    <ul data-list></ul>
-  </main>
-</body>
-
-</html>
